crypto_backend/requirements.txt
aiohttp==3.7.3
aiohttp-devtools==0.13.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
devtools==0.6.1
idna==2.10
multidict==5.1.0
numpy==1.19.4
Pillow==8.0.1
Pygments==2.7.3
tqdm==4.54.1
typing-extensions==3.7.4.3
watchgod==0.6
yarl==1.6.3
